Extract the type of FIT file

Quickly determine what type of FIT file you have without decoding all the data. Typically types include 'activity', 'workout' and 'course'. The full range of options is defined in the FIT file specification.

readFitFileType(fileName)

Arguments

fileName

A character specifying the FIT file to be read.

Value

As character vector of length 1. This will contain the text representation of the file type e.g. 'activity'

Examples

garmin_file <- system.file("extdata", "Activities", "garmin-edge530-ride.fit",
                           package = "FITfileR")
garmin <- readFitFileType(garmin_file)